transactional api
